08_Vue/08_vue_with_drf/vue_ws_8_3/blog_api_service/posts/serializers.py
from rest_framework import serializers
from .models import Post

class PostListSerializers(serializers.ModelSerializer):
    class Meta:
        model = Post
        fields = ('id', 'title', 'category',)