08_Vue/08_vue_with_drf/vue_ws_8_3/blog_api_service/requirements.txt
asgiref==3.8.1
Django==4.2.13
django-cors-headers==4.3.1
djangorestframework==3.15.1
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
